hello script
The script implements a retry mechanism to handle potential connection failures. The Test-Connection function is used to verify that the server is responding. If the server is unresponsive, the $serverStatus variable is set to $false.  If the server is unresponsive, the script retrieves the corresponding virtual machine in vCenter using the Get-VM function and the virtual machine name specified in the $vmName variable. Next, the Restart-VM function is used to perform a reset on the Virtua machine